stb_truetype.h: removed unnecesary lookups and declarations
These were found by the LLVM static analyzer. To the best of my
knowledge those values are never used. The test program still
works without them (which is perhaps not a good indication
as we might not be using a font with that format, better ask
Sean...). If those values were meant as documentation, perhaps
better to have them commented.

It wasn't going to make any difference for performance or suchlike,
any half-decent compiler would've eliminated those instructions.

...perhaps there was a reason for those statements to be there.